Pull DraftInputView, PostView into own files
